ai-playwright-framework/requirements.txt
playwright
pytest
pytest-html
allure-pytest
requests   # or httpx, depending on how you call Copilot API
pyyaml